+++ src/routes/ViewHoldingRoute.js
@@ -8,9 +8,16 @@ import { useStripes } from '@folio/stripes/core';
 
 import { useSearchInstanceByIdQuery } from '../common';
 import { DataContext } from '../contexts';
-import { useUpdateOwnership } from '../hooks';
+import {
+  useAuditSettings,
+  useUpdateOwnership,
+} from '../hooks';
 import ViewHoldingsRecord from '../ViewHoldingsRecord';
-import { UPDATE_OWNERSHIP_API } from '../constants';
+import {
+  INVENTORY_AUDIT_GROUP,
+  UPDATE_OWNERSHIP_API,
+} from '../constants';
+import { getIsVersionHistoryEnabled } from '../utils';
 
 const ViewHoldingRoute = () => {
   const { id: instanceId, holdingsrecordid } = useParams();
@@ -19,6 +26,9 @@ const ViewHoldingRoute = () => {
   const { state } = useLocation();
   const { instance } = useSearchInstanceByIdQuery(instanceId);
   const { updateOwnership } = useUpdateOwnership(UPDATE_OWNERSHIP_API.HOLDINGS);
+
+  const { settings } = useAuditSettings({ group: INVENTORY_AUDIT_GROUP });
+  const isVersionHistoryEnabled = getIsVersionHistoryEnabled(settings);
 
   return (
     <ViewHoldingsRecord
@@ -29,6 +39,7 @@ const ViewHoldingRoute = () => {
       referenceTables={referenceTables}
       holdingsrecordid={holdingsrecordid}
       onUpdateOwnership={updateOwnership}
+      isVersionHistoryEnabled={isVersionHistoryEnabled}
     />
   );
 };